Fixed bugs that were made visible by the equality-fixes in grammar
Repetitions parse into Arrays, not instances of Repetition. I’m still on the fence about this one, so I’ll keep the behavior for now. Regardless, now that grammar equality actually works, the existing test matcher is wrong. This nicely points out that I need to get this decision sorted so that things like this will be easier to get right.

The second part of the patch is yet another illustration of why it’s so import to get equality checking right. tsk tsk.