More fixes for handling optional Regexp embedded in other things
In this particular case the regex was the only element of a star-repeated Concatenation and the test case was trying to match the regex at the end of the input, which matched with an empty-string as the result. The any-repeat loop never terminated because the regex match succeeded (ie. didn’t return nil) on every loop.

A better solution would be to detect optional-regex’s and handle them more appropriately, but that’s not a simple thing to do in Ruby. For now it’s enough to simply break out of the repeat-loop at the end of the input.