Added RuleReference to EBNF
For keeping track of rule references when reading a grammar, particularly when one rule is an alias of another.

For example, in the grammar:

rule1 = "abc" | "xyz"
rule2 = rule1

Immediately resolving the reference would cause rule1 and rule2 to point to the same Grammar objects. For parsing purposes, this wouldn't be a problem in that a parser for the grammar would still work properly, however the resulting parse tree would potentially be ambiguous. Additionally, any attempt to export the grammar back to the original file format (or to any other) would produce a slightly different grammar.