Fixed handling of recursive references to identical rules
Loading an EBNF grammar containing references to recursive rules that happen to be identical, and writing that grammar back to EBNF, would cause the resulting rule references to become confused because Grammar::Recursion couldn't indicate which of the identical rules it was referencing. Added Parsers::RecursiveReference for disambiguating the references.